don't try and create archives when doing create_bin if no install directory exists

+++ scriptmodules/packages.sh
@@ -225,12 +225,16 @@ function rp_installBin() {
 
 function rp_createBin() {
     printHeading "Creating binary archive for $md_desc"
-    local archive="$md_id.tar.gz"
-    local dest="$__tmpdir/archives/$__platform/$md_type"
-    rm -f "$dest/$archive"
-    mkdir -p "$dest"
-    tar cvzf "$dest/$archive" -C "$rootdir/$md_type" "$md_id"
-    chown $user:$user "$dest/$archive"
+    if [[ -d "$rootdir/$md_type" ]]; then
+        local archive="$md_id.tar.gz"
+        local dest="$__tmpdir/archives/$__platform/$md_type"
+        rm -f "$dest/$archive"
+        mkdir -p "$dest"
+        tar cvzf "$dest/$archive" -C "$rootdir/$md_type" "$md_id"
+        chown $user:$user "$dest/$archive"
+    else
+        printMsgs "console" "No install directory $rootdir/$md_type - no archive created"
+    fi
 }
 
 function rp_registerModule() {
